+++ gapil/runtime/cc/map_test.cpp
@@ -21,7 +21,7 @@
 
 extern "C" {
 
-void* gapil_remap_pointer(context* ctx, uint64_t pointer, uint64_t length) {
+void* gapil_remap_pointer(context* ctx, uintptr_t pointer, uint64_t length) {
     return reinterpret_cast<void*>(pointer);
 }
