+++ gapil/runtime/cc/runtime.cpp
@@ -39,7 +39,7 @@ extern "C" {
 #include "runtime.h"
 
 // Declared externally
-void* gapil_remap_pointer(context* ctx, uint64_t pointer, uint64_t length);
+void* gapil_remap_pointer(context* ctx, uintptr_t pointer, uint64_t length);
 void  gapil_get_code_location(context* ctx, char** file, uint32_t* line);
 
 void gapil_init_context(context* ctx) {
@@ -154,7 +154,7 @@ void gapil_copy_slice(context* ctx, slice* dst, slice* src) {
     memcpy(dst->base, src->base, size);
 }
 
-void gapil_pointer_to_slice(context* ctx, uint64_t ptr, uint64_t offset, uint64_t size, slice* out) {
+void gapil_pointer_to_slice(context* ctx, uintptr_t ptr, uint64_t offset, uint64_t size, slice* out) {
     DEBUG_PRINT("gapil_pointer_to_slice(ptr: 0x%" PRIx64 ", offset: 0x%" PRIx64 ", size: 0x%" PRIx64 ")",
             ptr, offset, size);
 
@@ -168,7 +168,7 @@ void gapil_pointer_to_slice(context* ctx, uint64_t ptr, uint64_t offset, uint64_
     out->size = size;
 }
 
-string* gapil_pointer_to_string(context* ctx, uint64_t ptr) {
+string* gapil_pointer_to_string(context* ctx, uintptr_t ptr) {
     DEBUG_PRINT("gapil_pointer_to_string(ptr: 0x%" PRIx64 ")", ptr);
 
     auto data = reinterpret_cast<char*>(gapil_remap_pointer(ctx, ptr, 1));
